Moved dispatch functionality from AbstractActionController to AbstractController so AbstractController has default dispatch functions and removed duplication in AbstractRestfullController
