+++ library/Zend/Mvc/Controller/AbstractActionController.php
@@ -10,11 +10,8 @@
 
 namespace Zend\Mvc\Controller;
 
-use Zend\Http\PhpEnvironment\Response as HttpResponse;
 use Zend\Mvc\Exception;
 use Zend\Mvc\MvcEvent;
-use Zend\Stdlib\RequestInterface as Request;
-use Zend\Stdlib\ResponseInterface as Response;
 use Zend\View\Model\ViewModel;
 
 /**
@@ -60,38 +57,6 @@ abstract class AbstractActionController extends AbstractController
         return new ViewModel(array(
             'content' => 'Page not found'
         ));
-    }
-
-    /**
-     * Dispatch a request
-     *
-     * @events dispatch.pre, dispatch.post
-     * @param  Request $request
-     * @param  null|Response $response
-     * @return Response|mixed
-     */
-    public function dispatch(Request $request, Response $response = null)
-    {
-        $this->request = $request;
-        if (!$response) {
-            $response = new HttpResponse();
-        }
-        $this->response = $response;
-
-        $e = $this->getEvent();
-        $e->setRequest($request)
-          ->setResponse($response)
-          ->setTarget($this);
-
-        $result = $this->getEventManager()->trigger(MvcEvent::EVENT_DISPATCH, $e, function($test) {
-            return ($test instanceof Response);
-        });
-
-        if ($result->stopped()) {
-            return $result->last();
-        }
-
-        return $e->getResult();
     }
 
     /**
